RFC-257: Java Record Layer 4.14.2.0 migration, work to date
The migration branch's first commit gathers every workstream's uncommitted
work so far into one tree (owner: one branch for all migration code, one
workstream at a time from here).

State by workstream (rfcs/257-java-upgrade-audit/*-design.md):
- WS-A, WS-B: landed, reviewed.
- WS-C: revision 11 implemented; its review gate (graefe, torvalds,
  storage) returned Low findings that revision 12 folds next.
- WS-J: step 3 in progress. Landed here: the shape (d) refusal deleted
  under the owner's ruling that pre-release Go data is not supported;
  SetRecords reading the records file's extension options; NullStandin
  (NULL / NULL_UNIQUE / NOT_NULL) with Java's keyContainsNonUniqueNull at
  the unique checks and COUNT_NOT_NULL; a proto3 field at its default is
  absent to key evaluation and query reads (MessageHelpers
  .getFieldOnMessage); the key-function registry with the target's bounds,
  column sizes and nulls (FunctionSpec), create's refusals at Build and
  load, get_versionstamp_incarnation no longer a key function; an absent
  key-expression child is Java's refusal. Each is pinned against the live
  JVM in conformance. Open: IndexMaintenanceFilter (owner: port it).
- WS-D, WS-E, WS-F: designs in review, no implementation yet.
- Nightly RowDiff FDB watcher removed at the owner's request (TODO.md).

Review transcripts and tree freezes stay local
(rfcs/257-java-upgrade-audit/.gitignore); prompts and verdicts are
committed.

+++ pkg/recordlayer/index_predicate.go
@@ -343,7 +343,8 @@ func protoFieldToGoValue(m protoreflect.Message, fd protoreflect.FieldDescriptor
 	case protoreflect.Int64Kind, protoreflect.Sint64Kind, protoreflect.Sfixed64Kind:
 		return v.Int()
 	case protoreflect.Uint32Kind, protoreflect.Fixed32Kind:
-		return int64(v.Uint())
+		// Signed, as Java's protobuf-java reads a 32-bit unsigned field.
+		return int64(int32(uint32(v.Uint())))
 	case protoreflect.Uint64Kind, protoreflect.Fixed64Kind:
 		return int64(v.Uint())
 	case protoreflect.FloatKind:
